Move avatar crunch code to avatar-cruncher.js
